Consolidate records loading to the clubs store
